Fixes setuptools entry_points
Packaging is hard >.<
