+++ slack_invite_flow/email.plain.tmpl
