+++ tox.ini
@@ -2,5 +2,5 @@
 envlist=py27
 
 [testenv]
-commands=py.test slack-invite-flow/tests.py
+commands=py.test slack_invite_flow/tests.py
 deps=pytest
